Addition of a SpaceBlock Class
A spaceblock class has been added which has methods for manipulating the spaceblock properties.

An additional change has been made to the match method in the PuzzleBlockLogic class. The method name was changed and additionally, this method no longer returns a value. Instead, it assigns a boolean value to the is_matching_static_block