Directory definitions moved to config.py
The config file has been updated with constants pointing to the game's
data directory, a directory that will hold log statements and database
files. A function has also been created that checks if the data directory
exists in the HOMEPATH and creates one if it doesnt.

In the logger module, directory management called has been deleted and is
loaded instead from the config file.;